Refactored financial data types and centralized business logic to improve consistency, type safety, and code reusability.
Key changes include:
- Standardized walker pricing by changing `price` (int) to `hourlyRate` (double) across `WalkerModel`, `WalkerProfileScreen`, `BookingScreen`, and `ScheduleScreen`.
- Centralized the `checkAndResetEarnings` and `sendNotification` logic into `FirestoreService`, removing redundant implementations from UI screens.
- Enhanced `WalkerProfileEditScreen` with numeric input formatters and proper type casting for `hourlyRate` and `experienceYears`.
- Improved the chat system by extracting `MessageModel` into a standalone file and adding logic to increment a walker's `walksCount` upon session completion.
- Updated the owner-side UI by adding a "Messages" link to the drawer and refining navigation flows after booking completion.
- Refined UI data presentation by standardizing decimal formatting for ratings and hourly rates across the application.
- Improved code maintainability by cleaning up imports and moving business logic out of `build` methods.

+++ lib/screens/owner/booking_screen.dart
@@ -9,7 +9,7 @@ import '../../models/pet_model.dart';
 class BookingScreen extends StatefulWidget {
   final String walkerId;
   final String walkerName;
-  final int hourlyRate;
+  final double hourlyRate;
 
   const BookingScreen({
     super.key,
